+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/preload/BasePreloadManager.java
@@ -16,6 +16,7 @@
 package androidx.media3.exoplayer.source.preload;
 
 import static androidx.media3.common.util.Assertions.checkNotNull;
+import static androidx.media3.common.util.Util.postOrRun;
 
 import android.os.Handler;
 import android.os.Looper;
@@ -41,30 +42,30 @@ import java.util.PriorityQueue;
  * <p>Methods should be called on the same thread.
  */
 @UnstableApi
-public abstract class BasePreloadManager<T> {
+public abstract class BasePreloadManager<T, PreloadStatusT> {
 
   /** A base class of the builder of the concrete extension of {@link BasePreloadManager}. */
-  protected abstract static class BuilderBase<T> {
+  protected abstract static class BuilderBase<T, PreloadStatusT> {
 
     protected final Comparator<T> rankingDataComparator;
-    protected final TargetPreloadStatusControl<T> targetPreloadStatusControl;
+    protected final TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl;
     protected Supplier<MediaSource.Factory> mediaSourceFactorySupplier;
 
     public BuilderBase(
         Comparator<T> rankingDataComparator,
-        TargetPreloadStatusControl<T> targetPreloadStatusControl,
+        TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl,
         Supplier<MediaSource.Factory> mediaSourceFactorySupplier) {
       this.rankingDataComparator = rankingDataComparator;
       this.targetPreloadStatusControl = targetPreloadStatusControl;
       this.mediaSourceFactorySupplier = mediaSourceFactorySupplier;
     }
 
-    public abstract BasePreloadManager<T> build();
+    public abstract BasePreloadManager<T, PreloadStatusT> build();
   }
 
   private final Object lock;
   protected final Comparator<T> rankingDataComparator;
-  private final TargetPreloadStatusControl<T> targetPreloadStatusControl;
+  private final TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl;
   private final MediaSource.Factory mediaSourceFactory;
   private final ListenerSet<PreloadManagerListener> listeners;
   private final Map<MediaItem, MediaSourceHolder> mediaItemMediaSourceHolderMap;
@@ -75,11 +76,11 @@ public abstract class BasePreloadManager<T> {
 
   @GuardedBy("lock")
   @Nullable
-  private TargetPreloadStatusControl.PreloadStatus targetPreloadStatusOfCurrentPreloadingSource;
+  private PreloadStatusT targetPreloadStatusOfCurrentPreloadingSource;
 
   protected BasePreloadManager(
       Comparator<T> rankingDataComparator,
-      TargetPreloadStatusControl<T> targetPreloadStatusControl,
+      TargetPreloadStatusControl<T, PreloadStatusT> targetPreloadStatusControl,
       MediaSource.Factory mediaSourceFactory) {
     lock = new Object();
     applicationHandler = Util.createHandlerForCurrentOrMainLooper();
@@ -282,7 +283,7 @@ public abstract class BasePreloadManager<T> {
         return;
       }
     }
-    applicationHandler.post(() -> maybeAdvanceToNextSource(source));
+    postOrRun(applicationHandler, () -> maybeAdvanceToNextSource(source));
   }
 
   private void maybeAdvanceToNextSource(MediaSource currentSource) {
@@ -303,13 +304,9 @@ public abstract class BasePreloadManager<T> {
         && checkNotNull(sourceHolderPriorityQueue.peek()).mediaSource == mediaSource;
   }
 
-  /**
-   * Returns the {@linkplain TargetPreloadStatusControl.PreloadStatus target preload status} of the
-   * given {@link MediaSource}.
-   */
+  /** Returns the target preload status of the given {@link MediaSource}. */
   @Nullable
-  protected final TargetPreloadStatusControl.PreloadStatus getTargetPreloadStatus(
-      MediaSource source) {
+  protected final PreloadStatusT getTargetPreloadStatus(MediaSource source) {
     synchronized (lock) {
       if (!isPreloading(source)) {
         return null;
@@ -338,10 +335,11 @@ public abstract class BasePreloadManager<T> {
    * Preloads the given {@link MediaSource}.
    *
    * @param mediaSource The media source to preload.
-   * @param startPositionsUs The expected starting position in microseconds, or {@link C#TIME_UNSET}
-   *     to indicate the default start position.
+   * @param targetPreloadStatus The target preload status, may be null if a {@link MediaSource}
+   *     should not be preloaded.
    */
-  protected abstract void preloadSourceInternal(MediaSource mediaSource, long startPositionsUs);
+  protected abstract void preloadSourceInternal(
+      MediaSource mediaSource, @Nullable PreloadStatusT targetPreloadStatus);
 
   /**
    * Clears the preloaded data of the given {@link MediaSource}, while not releasing the instance of
@@ -362,8 +360,7 @@ public abstract class BasePreloadManager<T> {
   protected void releaseInternal() {}
 
   /**
-   * Starts to preload the {@link MediaSource} at the head of the priority queue, if the {@linkplain
-   * TargetPreloadStatusControl.PreloadStatus target preload status} for that source is not null.
+   * Starts to preload the {@link MediaSource} at the head of the priority queue.
    *
    * @return {@code true} if the {@link MediaSource} at the head of the priority queue starts to
    *     preload, otherwise {@code false}.
@@ -375,12 +372,9 @@ public abstract class BasePreloadManager<T> {
       MediaSourceHolder preloadingHolder = checkNotNull(sourceHolderPriorityQueue.peek());
       this.targetPreloadStatusOfCurrentPreloadingSource =
           targetPreloadStatusControl.getTargetPreloadStatus(preloadingHolder.rankingData);
-      if (targetPreloadStatusOfCurrentPreloadingSource != null) {
-        preloadSourceInternal(preloadingHolder.mediaSource, preloadingHolder.startPositionUs);
-        return true;
-      } else {
-        clearSourceInternal(preloadingHolder.mediaSource);
-      }
+      preloadSourceInternal(
+          preloadingHolder.mediaSource, targetPreloadStatusOfCurrentPreloadingSource);
+      return true;
     }
     return false;
   }
@@ -396,20 +390,14 @@ public abstract class BasePreloadManager<T> {
 
     public final MediaSource mediaSource;
     public final T rankingData;
-    public final long startPositionUs;
 
     public MediaSourceHolder(MediaSource mediaSource, T rankingData) {
-      this(mediaSource, rankingData, C.TIME_UNSET);
-    }
-
-    public MediaSourceHolder(MediaSource mediaSource, T rankingData, long startPositionUs) {
       this.mediaSource = mediaSource;
       this.rankingData = rankingData;
-      this.startPositionUs = startPositionUs;
     }
 
     @Override
-    public int compareTo(BasePreloadManager<T>.MediaSourceHolder o) {
+    public int compareTo(MediaSourceHolder o) {
       return rankingDataComparator.compare(this.rankingData, o.rankingData);
     }
   }